refactor alchemy
