+++ asapdiscovery-alchemy/asapdiscovery/alchemy/schema/prep_workflow.py
@@ -3,10 +3,10 @@ from typing import Any, Literal, Optional, Union
 
 import rich
 from asapdiscovery.alchemy.schema.base import _SchemaBase
-from asapdiscovery.data.schema_v2.complex import PreppedComplex
-from asapdiscovery.data.schema_v2.ligand import Ligand
-from asapdiscovery.data.state_expanders.protomer_expander import EpikExpander
-from asapdiscovery.data.state_expanders.stereo_expander import StereoExpander
+from asapdiscovery.data.schema.complex import PreppedComplex
+from asapdiscovery.data.schema.ligand import Ligand
+from asapdiscovery.data.structural.state_expanders.protomer_expander import EpikExpander
+from asapdiscovery.data.structural.state_expanders.stereo_expander import StereoExpander
 from asapdiscovery.docking.schema.pose_generation import (
     OpenEyeConstrainedPoseGenerator,
     RDKitConstrainedPoseGenerator,
